change variable type.
git-svn-id: https://realxtend-naali.googlecode.com/svn/trunk@2110 5b2332b8-efa3-11de-8684-7d64432d61a3

+++ CommunicationModule/TelepathyIM/FarsightChannel.h
@@ -64,7 +64,7 @@ namespace TelepathyIM
         GstPad *audio_in_src_pad_; // todo setter
         GstPad *video_in_src_pad_; // todo setter
         Status status_; // todo setter
-        gulong bus_watch_; // todo: setter
+        guint bus_watch_; // todo: setter
 
     private:
